+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [0.2.0] - (2021-05-09)
+
 ### Added
 
 - Add functions to change device options
@@ -24,5 +26,6 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 
 - Initial implementation of GET requests
 
-[unreleased]: https://github.com/dguo/awair-js/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/dguo/awair-js/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/dguo/awair-js/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/dguo/awair-js/releases/tag/v0.1.0